cloudsql-bigquery-backup-app/app/requirements.txt
google-api-python-client==2.97.0
google-auth==2.22.0
google-cloud-storage==2.10.0
oauth2client==4.1.3
